backend/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Background Tasks
celery[redis]==5.3.6

# Redis
redis==5.0.1

# AWS S3 (optional)
boto3==1.34.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Gemini
google-generativeai==0.3.2

# Video Processing
opencv-python==4.9.0.80
Pillow==10.2.0

# MQTT
paho-mqtt==1.6.1

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
